pyproject.toml
[build-system]
requires = ["setuptools>=69.5.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "direwolf-dashboard"
version = "0.1.0"
description = "Lightweight web-based live display of Direwolf activity for Raspberry Pi"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.11"
authors = [
    {name = "Craig Lamparter"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Communications :: Ham Radio",
]
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "aprslib>=0.7.0",
    "aprsd>=3.0.0",
    "aiosqlite>=0.19.0",
    "haversine>=2.8.0",
    "pyyaml>=6.0",
    "httpx>=0.25.0",
    "click>=8.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
]

[project.scripts]
direwolf-dashboard = "direwolf_dashboard.cli:main"

[project.urls]
Homepage = "https://github.com/hemna/direwolf-dashboard"
"Bug Reports" = "https://github.com/hemna/direwolf-dashboard/issues"
Source = "https://github.com/hemna/direwolf-dashboard"

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["direwolf_dashboard"]

[tool.setuptools.package-data]
direwolf_dashboard = [
    "static/**/*",
    "static/*",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
